Flytt hjelpefunksjoner for transformering av skjemadata ut i egen fil
