+++ src/frontend/sider/Fagsak/Dokumentutsending/DokumentutsendingContext.tsx
@@ -1,6 +1,7 @@
 import type { PropsWithChildren } from 'react';
 import { createContext, useContext, useEffect, useState } from 'react';
 
+import { transformerSkjemaData } from '@sider/Fagsak/Dokumentutsending/transformerSkjemaData';
 import deepEqual from 'deep-equal';
 
 import type { Avhengigheter, FeltState, ISkjema } from '@navikt/familie-skjema';
@@ -8,27 +9,22 @@ import { feil, ok, useFelt, useSkjema, Valideringsstatus } from '@navikt/familie
 import { type Ressurs, RessursStatus } from '@navikt/familie-typer';
 
 import { type DokumentÅrsak, DokumentÅrsakInstitusjon, DokumentÅrsakPerson } from './dokumentÅrsakTyper';
-import { hentEnkeltInformasjonsbrevRequest } from './Informasjonsbrev/enkeltInformasjonsbrevUtils';
 import useDokument from '../../../hooks/useDokument';
 import { Mottaker } from '../../../komponenter/Saklinje/Meny/LeggTilEllerFjernBrevmottakere/useBrevmottakerSkjema';
 import type { IManueltBrevRequestPåFagsak } from '../../../typer/dokument';
 import { Distribusjonskanal } from '../../../typer/dokument';
 import { FagsakType } from '../../../typer/fagsak';
-import type { IBarnMedOpplysninger } from '../../../typer/søknad';
-import { Målform } from '../../../typer/søknad';
+import type { IBarnMedOpplysninger, Målform } from '../../../typer/søknad';
 import { useBarnIBrevFelter } from '../../../utils/barnIBrevFelter';
 import type { IsoDatoString } from '../../../utils/dato';
-import { Datoformat, isoStringTilFormatertString } from '../../../utils/dato';
 import { useDeltBostedFelter } from '../../../utils/deltBostedSkjemaFelter';
 import type { IFritekstFelt } from '../../../utils/fritekstfelter';
 import { hentFrontendFeilmelding } from '../../../utils/ressursUtils';
-import type { ISelectOptionMedBrevtekst } from '../Behandling/Høyremeny/Brev/typer';
-import { Informasjonsbrev, opplysningsdokumenter } from '../Behandling/Høyremeny/Brev/typer';
 import { useBrukerContext } from '../BrukerContext';
 import { useFagsakContext } from '../FagsakContext';
 import { useManuelleBrevmottakerePåFagsakContext } from '../ManuelleBrevmottakerePåFagsakContext';
 
-interface DokumentutsendingSkjema {
+export interface DokumentutsendingSkjema {
     årsak: DokumentÅrsak | undefined;
     målform: Målform | undefined;
     fritekster: FeltState<IFritekstFelt>[];
@@ -214,151 +210,12 @@ export function DokumentutsendingProvider({ children }: PropsWithChildren) {
         nullstillSkjemaUtenomÅrsak();
     }, [årsak.verdi, bruker]);
 
-    const hentDeltBostedSkjemaData = (målform: Målform): IManueltBrevRequestPåFagsak => {
-        const barnIBrev = skjema.felter.barnMedDeltBosted.verdi.filter(barn => barn.merket);
-
-        return {
-            multiselectVerdier: barnIBrev.flatMap(hentDeltBostedMulitiselectVerdierForBarn),
-            barnIBrev: barnIBrev.map(barn => barn.ident),
-            mottakerMålform: målform,
-            brevmal: Informasjonsbrev.INFORMASJONSBREV_DELT_BOSTED,
-            manuelleBrevmottakere: manuelleBrevmottakerePåFagsak,
-        };
-    };
-
-    const hentBarnIBrevSkjemaData = (brevmal: Informasjonsbrev, målform: Målform): IManueltBrevRequestPåFagsak => {
-        const barnIBrev = skjema.felter.barnIBrev.verdi.filter(barn => barn.merket);
-
-        return {
-            multiselectVerdier: barnIBrev.map(
-                barn =>
-                    `Barn født ${isoStringTilFormatertString({
-                        isoString: barn.fødselsdato,
-                        tilFormat: Datoformat.DATO,
-                    })}.`
-            ),
-            barnIBrev: barnIBrev.map(barn => barn.ident),
-            mottakerMålform: målform,
-            brevmal: brevmal,
-            manuelleBrevmottakere: manuelleBrevmottakerePåFagsak,
-        };
-    };
-
-    const hentKanSøkeSkjemaData = (målform: Målform): IManueltBrevRequestPåFagsak => {
-        const fritekster = skjema.felter.fritekster.verdi.map(fritekstFelt => fritekstFelt.verdi.tekst);
-
-        const dokumenter = skjema.felter.dokumenter.verdi.map(dokumentOption => {
-            const dokument = opplysningsdokumenter.find(
-                dokument => dokument.label === dokumentOption
-            ) as ISelectOptionMedBrevtekst;
-            if (!dokument.brevtekst) {
-                throw new Error('Dokumentoptionen mangler brevtekst');
-            }
-            return dokument.brevtekst[målform];
+    const hentSkjemaData = (): IManueltBrevRequestPåFagsak =>
+        transformerSkjemaData({
+            skjema,
+            manuelleBrevmottakerePåFagsak,
+            hentDeltBostedMulitiselectVerdierForBarn,
         });
-
-        return {
-            multiselectVerdier: dokumenter.concat(fritekster),
-            barnIBrev: [],
-            mottakerMålform: målform,
-            brevmal: Informasjonsbrev.INFORMASJONSBREV_KAN_SØKE,
-            manuelleBrevmottakere: manuelleBrevmottakerePåFagsak,
-        };
-    };
-
-    const hentInnhenteOpplysningerKlageSkjemaData = (
-        brevmal: Informasjonsbrev,
-        målform: Målform
-    ): IManueltBrevRequestPåFagsak => {
-        return {
-            multiselectVerdier: [],
-            barnIBrev: [],
-            mottakerMålform: målform,
-            brevmal: brevmal,
-            manuelleBrevmottakere: manuelleBrevmottakerePåFagsak,
-            fritekstAvsnitt: fritekstAvsnitt.verdi,
-        };
-    };
-
-    const hentSkjemaData = (): IManueltBrevRequestPåFagsak => {
-        const dokumentÅrsak = skjema.felter.årsak.verdi;
-        if (dokumentÅrsak) {
-            switch (dokumentÅrsak) {
-                case DokumentÅrsakPerson.DELT_BOSTED:
-                    return hentDeltBostedSkjemaData(målform.verdi ?? Målform.NB);
-
-                case DokumentÅrsakPerson.FØDSEL_MINDREÅRIG:
-                    return hentEnkeltInformasjonsbrevRequest({
-                        målform: målform.verdi ?? Målform.NB,
-                        brevmal: Informasjonsbrev.INFORMASJONSBREV_FØDSEL_MINDREÅRIG,
-                        manuelleBrevmottakerePåFagsak,
-                    });
-                case DokumentÅrsakPerson.FØDSEL_VERGEMÅL:
-                    return hentEnkeltInformasjonsbrevRequest({
-                        målform: målform.verdi ?? Målform.NB,
-                        brevmal: Informasjonsbrev.INFORMASJONSBREV_FØDSEL_VERGEMÅL,
-                        manuelleBrevmottakerePåFagsak,
-                    });
-                case DokumentÅrsakPerson.FØDSEL_GENERELL:
-                    return hentEnkeltInformasjonsbrevRequest({
-                        målform: målform.verdi ?? Målform.NB,
-                        brevmal: Informasjonsbrev.INFORMASJONSBREV_FØDSEL_GENERELL,
-                        manuelleBrevmottakerePåFagsak,
-                    });
-                case DokumentÅrsakPerson.KAN_SØKE:
-                    return hentKanSøkeSkjemaData(målform.verdi ?? Målform.NB);
-                case DokumentÅrsakPerson.KAN_SØKE_EØS:
-                    return hentEnkeltInformasjonsbrevRequest({
-                        målform: målform.verdi ?? Målform.NB,
-                        brevmal: Informasjonsbrev.INFORMASJONSBREV_KAN_SØKE_EØS,
-                        manuelleBrevmottakerePåFagsak,
-                    });
-
-                case DokumentÅrsakPerson.TIL_FORELDER_MED_SELVSTENDIG_RETT_VI_HAR_FÅTT_F016_KAN_SØKE_OM_BARNETRYGD:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_TIL_FORELDER_MED_SELVSTENDIG_RETT_VI_HAR_FÅTT_F016_KAN_SØKE_OM_BARNETRYGD,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_GJORT_VEDTAK_TIL_ANNEN_FORELDER:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_GJORT_VEDTAK_TIL_ANNEN_FORELDER,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_ÅRLIG_KONTROLL:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_ÅRLIG_KONTROLL,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV:
-                    return hentBarnIBrevSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakPerson.INNHENTE_OPPLYSNINGER_KLAGE:
-                    return hentInnhenteOpplysningerKlageSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_INNHENTE_OPPLYSNINGER_KLAGE,
-                        målform.verdi ?? Målform.NB
-                    );
-                case DokumentÅrsakInstitusjon.INNHENTE_OPPLYSNINGER_KLAGE_INSTITUSJON:
-                    return hentInnhenteOpplysningerKlageSkjemaData(
-                        Informasjonsbrev.INFORMASJONSBREV_INNHENTE_OPPLYSNINGER_KLAGE_INSTITUSJON,
-                        målform.verdi ?? Målform.NB
-                    );
-            }
-        } else {
-            throw Error('Bruker ikke hentet inn og vi kan ikke sende inn skjema');
-        }
-    };
 
     const skjemaErLåst = () =>
         skjema.submitRessurs.status === RessursStatus.HENTER || hentetDokument.status === RessursStatus.HENTER;
